Initial setup
This sets up an initial rails app that has empty view and controller.

It uses docker and ruby on rails.  All of the project workflow is managed by shell scripts in the `script` folder. Checkout the `README.md` for some steps on getting started.

I've also added the first model to help us get started with the first table.